src/faker/generator.py
"""Core module for generating synthetic conversations."""

import os
import yaml
import logging
from typing import Any, Dict, List, Optional, Union

from faker.models import Conversation, Message, Dataset
from faker.llm import GeminiProvider

class ChatGenerator:
    """Main class for generating synthetic chat data.
    
    This class handles configuration, LLM interactions, and dataset generation.
    """
    
    def __init__(
        self, 
        llm_provider: Optional[Any] = None,
        templates: Optional[Dict[str, str]] = None,
        config: Optional[Dict[str, Any]] = None
    ):
        """Initialize a new chat generator.
        
        Args:
            llm_provider: The LLM provider to use for generation (default: GeminiProvider)
            templates: Dictionary of prompt templates for different generation scenarios
            config: Additional configuration parameters
        """
        self.llm = llm_provider or GeminiProvider()
        self.templates = templates or {}
        self.config = config or {}
        self.logger = logging.getLogger(__name__)
        
    @classmethod
    def from_config(cls, config_path: str) -> "ChatGenerator":
        """Create a new ChatGenerator from a configuration file.
        
        Args:
            config_path: Path to a YAML configuration file
            
        Returns:
            A configured ChatGenerator instance
        """
        with open(config_path, 'r') as f:
            config = yaml.safe_load(f)
            
        templates = config.get('templates', {})
        llm_config = config.get('llm', {})
        
        # Initialize the appropriate LLM provider
        provider_name = llm_config.get('provider', 'gemini')
        if provider_name == 'gemini':
            llm_provider = GeminiProvider(
                project_id=llm_config.get('project_id'),
                location=llm_config.get('location'),
                model=llm_config.get('model')
            )
        else:
            raise ValueError(f"Unsupported LLM provider: {provider_name}")
            
        return cls(
            llm_provider=llm_provider,
            templates=templates,
            config=config
        )
        
    def generate(self, num_conversations: int = 1) -> Dataset:
        """Generate a dataset of synthetic conversations.
        
        Args:
            num_conversations: Number of conversations to generate
            
        Returns:
            A Dataset object containing the generated conversations
        """
        conversations = []
        for i in range(num_conversations):
            self.logger.info(f"Generating conversation {i+1}/{num_conversations}")
            conversation = self._generate_conversation()
            conversations.append(conversation)
            
        return Dataset(conversations=conversations)
    
    def _generate_conversation(self) -> Conversation:
        """Generate a single conversation.
        
        Returns:
            A Conversation object containing messages
        """
        # Implementation details will depend on the specific generation approach
        # This is a placeholder that would be implemented based on the specific needs
        prompt = self.templates.get('conversation', 'Generate a realistic conversation between two people.')
        
        # Generate the conversation using the LLM
        result = self.llm.generate(prompt)
        
        # Parse the result into a structured conversation
        # This would be implemented based on the specific LLM response format
        messages = []
        # ... parsing logic here ...
        
        return Conversation(messages=messages)