pyproject.toml
[tool.poetry]
name = "faker"
version = "0.1.0"
description = "Modular synthetic chat data generation"
authors = ["Sean Koval"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-cloud-aiplatform = "^1.83.0"
pytest = "^8.3.5"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
